requirements.txt
numpy
scikit-learn
pandas
matplotlib
seaborn
feather-format
pyplotz
pyqt5
icecream
opencv-python-headless
